git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/struct-reorg-branch@91049 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libjava/javax/crypto/CipherInputStream.java
@@ -294,7 +294,7 @@ public class CipherInputStream extends FilterInputStream
    *
    * @param mark Is ignored.
    */
-  public void mark(int mark)
+  public void mark(long mark)
   {
   }
